Split out some modules into smaller ones (#2696)
* Rename AllocatePool to ObjectPool

* Split out various modules

* Lint fixes
